Proper exception handling using ControllerAdvice

+++ api/src/main/java/ca/me/proj/service/appointment/AppointmentService.java
@@ -1,12 +1,14 @@
 package ca.me.proj.service.appointment;
 
 import java.util.List;
+import javax.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import ca.me.proj.dtos.appointment.AppointmentDTO;
 import ca.me.proj.entity.appointment.AppointmentEntity;
 import ca.me.proj.entity.response.CustomResponseEntity;
+import ca.me.proj.exceptions.ResourceNotFoundException;
 import ca.me.proj.mapper.appointment.IAppointmentMapper;
 import ca.me.proj.repository.appointment.IAppointmentRepository;
 import ca.me.proj.repository.branch.IBranchRepository;
@@ -35,38 +37,45 @@ public class AppointmentService {
         return mapper.entityToDto(repository.findAll());
     }
 
-    public boolean existsById(Long id) {
+    public boolean existsById(long id) {
         return repository.existsById(id);
     }
 
-    public AppointmentDTO findById(Long id) {
+    public AppointmentDTO findById(long id) {
         return mapper.entityToDto(repository.findById(id).orElse(null));
     }
 
-    public ResponseEntity<String> createAppointment(AppointmentDTO dto) {
-        dto.setId(null);
-        AppointmentEntity entity = mapper.dtoToEntity(dto);
-        if (!branchRepository.existsById(entity.getBranchId())) {
-            return CustomResponseEntity.badRequestInvalidArgument("Branch ID does not exist");
-        } else if (!employeeRepository.existsById(entity.getEmployeeId())) {
-            return CustomResponseEntity.badRequestInvalidArgument("Employee ID does not exist");
-        } else if (!patientRepository.existsById(entity.getPatientId())) {
-            return CustomResponseEntity.badRequestInvalidArgument("Patient ID does not exist");
-        } else if (entity.getStartTime().after(entity.getEndTime())) {
-            return CustomResponseEntity.badRequestInvalidArgument("Start time is before end time");
-        } else if (repository.findPatientScheduleConflict(entity.getStartTime(),
-                entity.getEndTime(), entity.getPatientId())) {
-            return CustomResponseEntity.badRequestInvalidArgument("Patient has schedule conflict");
-        } else if (repository.findEmployeeScheduleConflict(entity.getStartTime(),
-                entity.getEndTime(), entity.getEmployeeId())) {
-            return CustomResponseEntity.badRequestInvalidArgument("Employee has schedule conflict");
-        } else {
-            repository.save(entity);
-            return CustomResponseEntity.saveSuccess();
-        }
+    public ResponseEntity<String> createAppointment(@Valid AppointmentDTO dto) {
+
+        repository.save(mapper.dtoToEntity(dto));
+        return CustomResponseEntity.saveSuccess();
+        // dto.setId(null);
+        // AppointmentEntity entity = mapper.dtoToEntity(dto);
+        // if (!branchRepository.existsById(entity.getBranchId())) {
+        // return CustomResponseEntity.badRequestInvalidArgument("Branch ID does not exist");
+        // } else if (!employeeRepository.existsById(entity.getEmployeeId())) {
+        // return CustomResponseEntity.badRequestInvalidArgument("Employee ID does not exist");
+        // } else if (!patientRepository.existsById(entity.getPatientId())) {
+        // return CustomResponseEntity.badRequestInvalidArgument("Patient ID does not exist");
+        // } else if (entity.getStartTime().after(entity.getEndTime())) {
+        // return CustomResponseEntity.badRequestInvalidArgument("Start time is before end time");
+        // } else if (repository.findPatientScheduleConflict(entity.getStartTime(),
+        // entity.getEndTime(), entity.getPatientId())) {
+        // return CustomResponseEntity.badRequestInvalidArgument("Patient has schedule conflict");
+        // } else if (repository.findEmployeeScheduleConflict(entity.getStartTime(),
+        // entity.getEndTime(), entity.getEmployeeId())) {
+        // return CustomResponseEntity.badRequestInvalidArgument("Employee has schedule conflict");
+        // } else {
+        // repository.save(entity);
+        // return CustomResponseEntity.saveSuccess();
+        // }
     }
 
-    public ResponseEntity<String> deletebyId(Long id) {
+    public ResponseEntity<String> deletebyId(long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("An entity with id " + id + " does not exist");
+        }
+
         if (repository.existsById(id)) {
             repository.deleteById(id);
             return CustomResponseEntity.deleteSuccess();
